# internal/api/v1/docs/swagger.yaml
definitions:
  Agent:
    properties:
      addr:
        type: string
      name:
        type: string
      version:
        type: string
    type: object
  CIDR:
    properties:
      ip:
        description: network number
        items:
          type: integer
        type: array
      mask:
        description: network mask
        items:
          type: integer
        type: array
    type: object
...
  Container:
    properties:
      agent:
        $ref: '#/definitions/Agent'
      aliases:
        items:
          type: string
        type: array
      container_id:
        type: string
      container_name:
        type: string
      docker_host:
        type: string
      errors:
        type: string
      idlewatcher_config:
        $ref: '#/definitions/IdlewatcherConfig'
      image:
        $ref: '#/definitions/ContainerImage'
      is_excluded:
        type: boolean
      is_explicit:
        type: boolean
      is_host_network_mode:
        type: boolean
      mounts:
        items:
          type: string
        type: array
      network:
        type: string
      private_hostname:
        type: string
      private_ports:
        allOf:
        - $ref: '#/definitions/types.PortMapping'
        description: privatePort:types.Port
      public_hostname:
        type: string
      public_ports:
        allOf:
        - $ref: '#/definitions/types.PortMapping'
        description: non-zero publicPort:types.Port
      running:
        type: boolean
    type: object
  ContainerImage:
    properties:
      author:
        type: string
      name:
        type: string
      tag:
        type: string
    type: object
...
  ContainerStopMethod:
    enum:
    - pause
    - stop
    - kill
    type: string
    x-enum-varnames:
    - ContainerStopMethodPause
    - ContainerStopMethodStop
    - ContainerStopMethodKill
  DockerConfig:
    properties:
      container_id:
        type: string
      container_name:
        type: string
      docker_host:
        type: string
    required:
    - container_id
    - container_name
    - docker_host
    type: object
  ErrorResponse:
    properties:
      error:
        type: string
        x-nullable: true
      message:
        type: string
    type: object
...
  HTTPHeader:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  HealthCheckConfig:
    properties:
      disable:
        type: boolean
      interval:
        type: integer
      path:
        type: string
      retries:
        description: '<0: immediate, >=0: threshold'
        type: integer
      timeout:
        type: integer
      use_get:
        type: boolean
    type: object
  HealthExtra:
    properties:
      config:
        $ref: '#/definitions/LoadBalancerConfig'
      pool:
        additionalProperties: {}
        type: object
    type: object
  HealthJSON:
    properties:
      config:
        $ref: '#/definitions/HealthCheckConfig'
      detail:
        type: string
      extra:
        allOf:
        - $ref: '#/definitions/HealthExtra'
        x-nullable: true
      lastSeen:
        type: integer
      lastSeenStr:
        type: string
      latency:
        type: number
      latencyStr:
        type: string
      name:
        type: string
      started:
        type: integer
      startedStr:
        type: string
      status:
        type: string
      uptime:
        type: number
      uptimeStr:
        type: string
      url:
        type: string
    type: object
...
  HomepageOverrideCategoryOrderParams:
    properties:
      value:
        type: integer
      which:
        type: string
    type: object
  HomepageOverrideItemParams:
    properties:
      value:
        $ref: '#/definitions/homepage.ItemConfig'
      which:
        type: string
    type: object
  HomepageOverrideItemVisibleParams:
    properties:
      value:
        type: boolean
      which:
        items:
          type: string
        type: array
    type: object
  HomepageOverrideItemsBatchParams:
    properties:
      value:
        additionalProperties:
          $ref: '#/definitions/homepage.ItemConfig'
        type: object
    type: object
  IdlewatcherConfig:
    properties:
      depends_on:
        items:
          type: string
        type: array
      docker:
        $ref: '#/definitions/DockerConfig'
      idle_timeout:
        allOf:
        - $ref: '#/definitions/time.Duration'
        description: "0: no idle watcher.\nPositive: idle watcher with idle timeout.\nNegative:
          idle watcher as a dependency.\tIdleTimeout time.Duration `json:\"idle_timeout\"
          json_ext:\"duration\"`"
      proxmox:
        $ref: '#/definitions/ProxmoxConfig'
      start_endpoint:
        description: Optional path that must be hit to start container
        type: string
      stop_method:
        $ref: '#/definitions/ContainerStopMethod'
      stop_signal:
        type: string
      stop_timeout:
        $ref: '#/definitions/time.Duration'
      wake_timeout:
        $ref: '#/definitions/time.Duration'
    type: object
  ListFilesResponse:
    properties:
      config:
        items:
          type: string
        type: array
      middleware:
        items:
          type: string
        type: array
      provider:
        items:
          type: string
        type: array
    type: object
  LoadBalancerConfig:
    properties:
      link:
        type: string
      mode:
        $ref: '#/definitions/LoadBalancerMode'
      options:
        additionalProperties: {}
        type: object
      weight:
        type: integer
    type: object
  LoadBalancerMode:
    enum:
    - ""
    - roundrobin
    - leastconn
    - iphash
    type: string
    x-enum-varnames:
    - LoadbalanceModeUnset
    - LoadbalanceModeRoundRobin
    - LoadbalanceModeLeastConn
    - LoadbalanceModeIPHash
  LogFilter-CIDR:
    properties:
      negative:
        type: boolean
      values:
        items:
          $ref: '#/definitions/CIDR'
        type: array
    type: object
  LogFilter-HTTPHeader:
    properties:
      negative:
        type: boolean
      values:
        items:
          $ref: '#/definitions/HTTPHeader'
        type: array
    type: object
  LogFilter-HTTPMethod:
    properties:
      negative:
        type: boolean
      values:
        items:
          type: string
        type: array
    type: object
  LogFilter-Host:
    properties:
      negative:
        type: boolean
      values:
        items:
          type: string
        type: array
    type: object
  LogFilter-StatusCodeRange:
    properties:
      negative:
        type: boolean
      values:
        items:
          $ref: '#/definitions/StatusCodeRange'
        type: array
    type: object
  LogRetention:
    properties:
      days:
        type: integer
      keep_size:
        type: integer
      last:
        type: integer
    type: object
...
  ProxmoxConfig:
    properties:
      node:
        type: string
      vmid:
        type: integer
    required:
    - node
    - vmid
    type: object
...
  RequestLoggerConfig:
    properties:
      buffer_size:
        description: 'Deprecated: buffer size is adjusted dynamically'
        type: integer
      fields:
        $ref: '#/definitions/accesslog.Fields'
      filters:
        $ref: '#/definitions/accesslog.Filters'
      format:
        enum:
        - common
        - combined
        - json
        type: string
      path:
        type: string
      retention:
        $ref: '#/definitions/LogRetention'
      rotate_interval:
        type: integer
      stdout:
        type: boolean
    type: object
  Route:
    properties:
      access_log:
        allOf:
        - $ref: '#/definitions/RequestLoggerConfig'
        x-nullable: true
      agent:
        type: string
      alias:
        type: string
      container:
        allOf:
        - $ref: '#/definitions/Container'
        description: Docker only
        x-nullable: true
      disable_compression:
        type: boolean
      excluded:
        type: boolean
      health:
        allOf:
        - $ref: '#/definitions/HealthJSON'
        description: for swagger
      healthcheck:
        $ref: '#/definitions/HealthCheckConfig'
      homepage:
        $ref: '#/definitions/homepage.ItemConfig'
      host:
        type: string
      idlewatcher:
        allOf:
        - $ref: '#/definitions/IdlewatcherConfig'
        x-nullable: true
      load_balance:
        allOf:
        - $ref: '#/definitions/LoadBalancerConfig'
        x-nullable: true
      lurl:
        description: private fields
        type: string
        x-nullable: true
      middlewares:
        additionalProperties:
          $ref: '#/definitions/types.LabelMap'
        type: object
        x-nullable: true
      no_tls_verify:
        type: boolean
      path_patterns:
        items:
          type: string
        type: array
        x-nullable: true
      port:
        $ref: '#/definitions/route.Port'
      provider:
        description: for backward compatibility
        type: string
        x-nullable: true
      purl:
        type: string
      response_header_timeout:
        type: integer
      root:
        type: string
      rules:
        items:
          $ref: '#/definitions/rules.Rule'
        type: array
        uniqueItems: true
      scheme:
        $ref: '#/definitions/route.Scheme'
    type: object
...
  StatusCodeRange:
    properties:
      end:
        type: integer
      start:
        type: integer
    type: object
  SuccessResponse:
    properties:
      details:
        additionalProperties: {}
        type: object
        x-nullable: true
      message:
        type: string
    type: object
...
  accesslog.FieldConfig:
    properties:
      config:
        additionalProperties:
          $ref: '#/definitions/accesslog.FieldMode'
        type: object
      default:
        allOf:
        - $ref: '#/definitions/accesslog.FieldMode'
        enum:
        - keep
        - drop
        - redact
    type: object
  accesslog.FieldMode:
    enum:
    - keep
    - drop
    - redact
    type: string
    x-enum-varnames:
    - FieldModeKeep
    - FieldModeDrop
    - FieldModeRedact
  accesslog.Fields:
    properties:
      cookies:
        $ref: '#/definitions/accesslog.FieldConfig'
      headers:
        $ref: '#/definitions/accesslog.FieldConfig'
      query:
        $ref: '#/definitions/accesslog.FieldConfig'
    type: object
  accesslog.Filters:
    properties:
      cidr:
        $ref: '#/definitions/LogFilter-CIDR'
      headers:
        allOf:
        - $ref: '#/definitions/LogFilter-HTTPHeader'
        description: header exists or header == value
      host:
        $ref: '#/definitions/LogFilter-Host'
      method:
        $ref: '#/definitions/LogFilter-HTTPMethod'
      status_codes:
        $ref: '#/definitions/LogFilter-StatusCodeRange'
    type: object
...
  container.Port:
    properties:
      IP:
        description: Host IP address that the container's port is mapped to
        type: string
      PrivatePort:
        description: |-
          Port on the container
          Required: true
        type: integer
      PublicPort:
        description: Port exposed on the host
        type: integer
      Type:
        description: |-
          type
          Required: true
        type: string
    type: object
...
  homepage.IconExtra:
    properties:
      file_type:
        type: string
      is_dark:
        type: boolean
      is_light:
        type: boolean
      key:
        type: string
      ref:
        type: string
    type: object
...
  homepage.IconSource:
    enum:
    - https://
    - '@target'
    - '@walkxcode'
    - '@selfhst'
    type: string
    x-enum-varnames:
    - IconSourceAbsolute
    - IconSourceRelative
    - IconSourceWalkXCode
    - IconSourceSelfhSt
  homepage.IconURL:
    properties:
      extra:
        allOf:
        - $ref: '#/definitions/homepage.IconExtra'
        description: only for walkxcode/selfhst icons
      source:
        $ref: '#/definitions/homepage.IconSource'
      value:
        description: only for absolute/relative icons
        type: string
    type: object
...
  homepage.ItemConfig:
    properties:
      category:
        type: string
      description:
        type: string
      icon:
        $ref: '#/definitions/homepage.IconURL'
      name:
        description: display name
        type: string
      show:
        type: boolean
      sort_order:
        type: integer
      url:
        type: string
    type: object
  homepageapi.SetHomePageOverridesRequest:
    properties:
      value: {}
      what:
        enum:
        - item
        - items_batch
        - category_order
        - item_visible
        type: string
    required:
    - value
    - what
    type: object
...
  route.Port:
    properties:
      listening:
        type: integer
      proxy:
        type: integer
    type: object
...
  route.Scheme:
    enum:
    - http
    - https
    - tcp
    - udp
    - fileserver
    type: string
    x-enum-varnames:
    - SchemeHTTP
    - SchemeHTTPS
    - SchemeTCP
    - SchemeUDP
    - SchemeFileServer
...
  rules.Command:
    type: object
  rules.Rule:
    properties:
      do:
        $ref: '#/definitions/rules.Command'
      name:
        type: string
      "on":
        $ref: '#/definitions/rules.RuleOn'
    type: object
  rules.RuleOn:
    type: object
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
  types.LabelMap:
    additionalProperties: {}
    type: object
  types.PortMapping:
    additionalProperties:
      $ref: '#/definitions/container.Port'
    type: object
...
info:
  contact: {}
paths:
...
  /agent/list:
    get:
      consumes:
      - application/json
      description: List agents
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Agent'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List agents
      tags:
      - agent
      - websocket
      x-id: list
...
  /file/list:
    get:
      consumes:
      - application/json
      description: List files
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ListFilesResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List files
      tags:
      - file
      x-id: list
  /file/validate:
    post:
      consumes:
      - application/json
      description: Validate file
      parameters:
      - description: Type
        enum:
        - config
        - provider
        - middleware
        in: query
        name: type
        required: true
        type: string
      - description: File content
        in: body
        name: file
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: File validated
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "417":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Validate file
      tags:
      - file
      x-id: validate
...
  /homepage/set:
    post:
      consumes:
      - application/json
      description: Set homepage overrides
      parameters:
      - description: Override single item
        in: body
        name: request
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/homepageapi.SetHomePageOverridesRequest'
          - properties:
              value:
                $ref: '#/definitions/HomepageOverrideItemParams'
            type: object
      - description: Override multiple items
        in: body
        name: request
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/homepageapi.SetHomePageOverridesRequest'
          - properties:
              value:
                $ref: '#/definitions/HomepageOverrideItemsBatchParams'
            type: object
      - description: Override category order
        in: body
        name: request
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/homepageapi.SetHomePageOverridesRequest'
          - properties:
              value:
                $ref: '#/definitions/HomepageOverrideCategoryOrderParams'
            type: object
      - description: Override item visibility
        in: body
        name: request
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/homepageapi.SetHomePageOverridesRequest'
          - properties:
              value:
                $ref: '#/definitions/HomepageOverrideItemVisibleParams'
            type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Set homepage overrides
      tags:
      - homepage
      x-id: set
...
  /route/{which}:
    get:
      consumes:
      - application/json
      description: List route
      parameters:
      - description: Route name
        in: path
        name: which
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Route'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List route
      tags:
      - route
      x-id: route
...
  /route/list:
    get:
      consumes:
      - application/json
      description: List routes
      parameters:
      - description: Provider
        in: query
        name: provider
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Route'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List routes
      tags:
      - route
      - websocket
      x-id: routes
...
swagger: "2.0"